+++ src/commonMain/kotlin/csense/kotlin/extensions/Exceptions.kt
@@ -15,10 +15,12 @@ import csense.kotlin.logger.LoggingFunctionType
  * @param throwableAction EmptyFunctionResult<T>
  * @return T?
  */
-inline fun <T> tryAndLog(title: String = "",
-                         message: String = "",
-                         logger: LoggingFunctionType<*> = L::error,
-                         throwableAction: EmptyFunctionResult<T>): T? {
+inline fun <T> tryAndLog(
+        title: String = "",
+        message: String = "",
+        logger: LoggingFunctionType<*> = L::error,
+        throwableAction: EmptyFunctionResult<T>
+): T? {
     return try {
         throwableAction()
     } catch (e: Throwable) {