+++ src/commonTest/kotlin/csense/kotlin/extensions/collections/map/MapKtTest.kt
@@ -2,16 +2,15 @@
 
 package csense.kotlin.extensions.collections.map
 
-import csense.kotlin.tests.assertions.assert
-import csense.kotlin.tests.assertions.failTest
+import csense.kotlin.tests.assertions.*
 import kotlin.test.Test
 
 
 class MapKtTest {
-
+    
     val emptyMap = mapOf<String, String>()
     val singleMap = mapOf("a" to "b")
-
+    
     @Test
     fun forEachIndexed() {
         emptyMap.forEachIndexed { _, _: Int ->
@@ -23,7 +22,7 @@ class MapKtTest {
             index.assert(0)
         }
     }
-
+    
     @Test
     fun filterMapKey() {
         emptyMap.forEachIndexed { _, _ ->
@@ -35,7 +34,7 @@ class MapKtTest {
             index.assert(0)
         }
     }
-
+    
     @Test
     fun useValueOr() {
         var notFoundCounter = 0
@@ -48,6 +47,32 @@ class MapKtTest {
         singleMap.useValueOr("b", { failTest() }, { notFoundCounter += 1 })
         notFoundCounter.assert(1)
     }
-
+    
+    class MapKVDoesNotContainKey() {
+        @Test
+        fun empty() {
+            val map = mapOf<String, String>()
+            map.doesNotContainKey("1").assertTrue()
+        }
+        
+        @Test
+        fun single() {
+            val map = mapOf("1234" to "")
+            map.doesNotContainKey("1").assertTrue()
+            map.doesNotContainKey("1234").assertFalse()
+            map.doesNotContainKey("12345").assertTrue()
+        }
+        
+        @Test
+        fun multiple() {
+            val map = mapOf(123 to "", 444 to "")
+            map.doesNotContainKey(1).assertTrue()
+            map.doesNotContainKey(2).assertTrue()
+            map.doesNotContainKey(123).assertFalse()
+            map.doesNotContainKey(444).assertFalse()
+            map.doesNotContainKey(555).assertTrue()
+        }
+        
+    }
 }